Adapt to the removal of /usr/bin/package-manager
libzypp has removed the /usr/bin/package-manager script, so just call
YaST's sw_single module. Calling yast2 --install does not make sense
because it is deprecated as well, thus package-manager.desktop can no
longer be a MIME provider for rpms and just serves as a shortcut to
YaST's package installer on desktops which do not have a PackageKit
frontend.